- move options modules under ./lib/cli directory
