+++ lib/cli/clirunner.js
@@ -6,7 +6,7 @@ const CodeRunner    = require('../index'),
       terminal      = require('../util/terminal'),
       packageConfig = require(`${__dirname}/../../package.json`);
 
-const getRunOptions = require('../options');
+const getRunOptions = require('./options');
 
 process.on('unhandledRejection', r => logger.error(r));
