+++ lib/cli/consul.js
