+++ lib/cli/env.js
