+++ lib/cli/file.js
